# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "MeshCore network observation backend. Ingests LoRa packets from MQTT brokers, stores in PostgreSQL, and streams live events via WebSocket.",
        "title": "MeshCore Beacon API",
        "termsOfService": "https://github.com/MeshCore-Beacon/beacon-server",
        "contact": {
            "name": "MeshCore Beacon",
            "url": "https://github.com/MeshCore-Beacon/beacon-server"
        },
        "license": {
            "name": "AGPL-3-or-later"
        },
        "version": "1.4.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/atlas/regions/{slug}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Atlas"
                ],
                "summary": "Get regional Atlas story summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Region slug, or all",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Window start as epoch milliseconds",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Window end as epoch milliseconds",
                        "name": "until",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RegionAtlasSummary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
...
        "/iatas": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IATAs"
                ],
                "summary": "List all IATA codes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.IATA"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/iatas/{iata}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IATAs"
                ],
                "summary": "Get a single IATA code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "3-letter IATA code",
                        "name": "iata",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.IATA"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
...
        "/regions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Regions"
                ],
                "summary": "List all regions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RegionSummary"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/regions/{regionId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Regions"
                ],
                "summary": "Get a single region",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Region ID",
                        "name": "regionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Region"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
...
        "/stats/regions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stats"
                ],
                "summary": "Stats per-IATA regional comparison",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated IATA codes",
                        "name": "iatas",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by region ID, expands to member IATAs",
                        "name": "regionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by region slug, expands to member IATAs",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Window preset: 24h, 7d, or 30d",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Window start as epoch milliseconds",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Window end as epoch milliseconds",
                        "name": "until",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bucket size: 1h, 6h, or 24h",
                        "name": "bucket",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.StatsRegions"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.AtlasIATA": {
            "type": "object",
            "properties": {
                "activeObservers": {
                    "type": "integer"
                },
                "displayName": {
                    "type": "string"
                },
                "iata": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "observationCount": {
                    "type": "integer"
                },
                "uniquePackets": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.AtlasStoryBeat": {
            "type": "object",
            "properties": {
                "at": {
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "iata": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.AtlasWindow": {
            "type": "object",
            "properties": {
                "since": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "until": {
                    "description": "epoch ms",
                    "type": "integer"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.IATA": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "iata": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.NodeTypeCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "nodeType": {
                    "type": "integer"
                },
                "nodeTypeName": {
                    "type": "string"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.ObservationPoint": {
            "type": "object",
            "properties": {
                "activeObservers": {
                    "type": "integer"
                },
                "hour": {
                    "description": "epoch ms, start of the 1-hour bucket",
                    "type": "integer"
                },
                "iata": {
                    "type": "string"
                },
                "observationCount": {
                    "type": "integer"
                },
                "uniquePackets": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.PayloadBreakdownItem": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "payloadType": {
                    "type": "integer"
                },
                "payloadTypeName": {
                    "type": "string"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.RadioPreset": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "number of observers or nodes on this preset in this IATA",
                    "type": "integer"
                },
                "iata": {
                    "type": "string"
                },
                "preset": {
                    "description": "\"freqMhz,bwKhz,sf\" e.g. \"910.525,62.5,7\"",
                    "type": "string"
                },
                "sourceType": {
                    "description": "\"observer\" or \"node\"",
                    "type": "string"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.Region": {
            "type": "object",
            "properties": {
                "centerLat": {
                    "description": "map center latitude",
                    "type": "number"
                },
                "centerLng": {
                    "description": "map center longitude",
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "iatas": {
                    "description": "member IATA codes",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "description": "URL-safe identifier e.g. \"western-canada\"",
                    "type": "string"
                },
                "zoomLevel": {
                    "description": "suggested map zoom level",
                    "type": "integer"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.RegionAtlasSummary": {
            "type": "object",
            "properties": {
                "hourly": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObservationPoint"
                    }
                },
                "iatas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.AtlasIATA"
                    }
                },
                "kpis": {
                    "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.StatsOverview"
                },
                "nodeTypes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.NodeTypeCount"
                    }
                },
                "payloadMix": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PayloadBreakdownItem"
                    }
                },
                "radioPresets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.RadioPreset"
                    }
                },
                "region": {
                    "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Region"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ScopeSummary"
                    }
                },
                "storyBeats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.AtlasStoryBeat"
                    }
                },
                "topNodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.TopNode"
                    }
                },
                "topObservers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.TopObserver"
                    }
                },
                "window": {
                    "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.AtlasWindow"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.RegionSummary": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "description": "URL-safe identifier e.g. \"western-canada\"",
                    "type": "string"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.ScopeSummary": {
            "type": "object",
            "properties": {
                "iataCount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nodeCount": {
                    "type": "integer"
                },
                "observerCount": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.StatsOverview": {
            "type": "object",
            "properties": {
                "activeIatas": {
                    "type": "integer"
                },
                "activeObservers": {
                    "type": "integer"
                },
                "totalObservations": {
                    "type": "integer"
                },
                "totalPackets": {
                    "type": "integer"
                },
                "windowHours": {
                    "description": "always 24 for now",
                    "type": "integer"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.StatsRegionRow": {
            "type": "object",
            "properties": {
                "activeNodes": {
                    "type": "integer"
                },
                "activeObservers": {
                    "type": "integer"
                },
                "iata": {
                    "type": "string"
                },
                "lastHeard": {
                    "type": "integer"
                },
                "observationCount": {
                    "type": "integer"
                },
                "packetCount": {
                    "type": "integer"
                },
                "topPayloadCount": {
                    "type": "integer"
                },
                "topPayloadType": {
                    "type": "integer"
                },
                "topPayloadTypeName": {
                    "type": "string"
                },
                "topRouteCount": {
                    "type": "integer"
                },
                "topRouteType": {
                    "type": "integer"
                },
                "topRouteTypeName": {
                    "type": "string"
                },
                "trend": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.StatsTrendPoint"
                    }
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.StatsRegions": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.StatsRegionRow"
                    }
                },
                "serverTime": {
                    "type": "integer"
                },
                "window": {
                    "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.StatsWindow"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.StatsTrendPoint": {
            "type": "object",
            "properties": {
                "activeObservers": {
                    "type": "integer"
                },
                "observationCount": {
                    "type": "integer"
                },
                "packetCount": {
                    "type": "integer"
                },
                "t": {
                    "type": "integer"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.StatsWindow": {
            "type": "object",
            "properties": {
                "bucket": {
                    "type": "string"
                },
                "since": {
                    "type": "integer"
                },
                "until": {
                    "type": "integer"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.TopNode": {
            "type": "object",
            "properties": {
                "iata": {
                    "type": "string"
                },
                "lastHeard": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "nodeId": {
                    "type": "string"
                },
                "nodeName": {
                    "type": "string"
                },
                "nodeType": {
                    "type": "integer"
                },
                "nodeTypeName": {
                    "type": "string"
                },
                "observationCount": {
                    "type": "integer"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.TopObserver": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "iata": {
                    "type": "string"
                },
                "observationCount": {
                    "type": "integer"
                },
                "observerId": {
                    "type": "string"
                },
                "observerType": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g. \"not_found\", \"bad_request\"",
                    "type": "string"
                },
                "message": {
                    "description": "e.g. \"channel not found\"",
                    "type": "string"
                }
            }
        },
...
    },
...
}